requirements.txt
Django==2.1.8
mysqlclient==1.4.2.post1
PyMySQL
django-import-export
django-admin-numeric-filter
djangorestframework
djangorestframework-datatables
htrc-feature-reader
